app/requirements.txt
aiogramaiogram==2.22.2
python-dotenv
aiohttp